ghidra_jupyter/requirements.txt
psutil
attrs
notebook
tqdm
requests